refactor: reorganiza pastas por feature e divide service de escala
