+++ frontend-app/src/components/layout/Header.tsx
@@ -3,7 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import { Button } from "../ui/Button";
 import { PlusCircle } from "lucide-react";
 import { useState } from "react";
-import { ModalServico } from "../pages/Dashboard/ModalServico";
+import { ModalServico } from "../../features/dashboard/components/ModalServico";
 
 export const Header = () => {
 //  vamos puxar da API usando o react Query
