+++ frontend-app/src/features/dashboard/components/ModalServico.tsx
@@ -1,7 +1,7 @@
 import { Plus, Trash2, X } from "lucide-react"
-import { Button } from "../../ui/Button"
-import { Input } from "../../ui/Input"
-import { useModalServico } from "../../../hooks/useModalServico"
+import { Button } from "../../../components/ui/Button"
+import { Input } from "../../../components/ui/Input"
+import { useModalServico } from "../hooks/useModalServico"
 import type { Aluno } from "@prisma/client"
 
 interface ModalProps {
