+++ frontend-app/src/features/dashboard/services/servicos.service.ts
@@ -1,4 +1,4 @@
-import { api } from "./api";
+import { api } from "../../../services/api";
 
 interface PayloadNovoServico {
   data: string;
